refactor(list): update formatting and improve list circular dependencies
