+++ deno_dist/list/list-custom.ts
@@ -4,6 +4,7 @@ export * from './implementation/nonleaf/interface.ts';
 export * from './implementation/nonleaf/nonleaf-block.ts';
 export * from './implementation/tree/interface.ts';
 export * from './implementation/tree/operations.ts';
+export * from './implementation/leaf/non-empty.ts';
 
 export * from './builder/block-builder.ts';
 export * from './builder/builder-base.ts';
@@ -11,15 +12,15 @@ export * from './builder/gen-builder.ts';
 export * from './builder/leaf/block-builder.ts';
 export * from './builder/leaf/leaf-builder.ts';
 export * from './builder/nonleaf/nonleaf-builder.ts';
+export * from './builder/tree/tree-builder.ts';
+export * from './builder/nonleaf/block-builder.ts';
 
 // circular dependencies
-export * from './implementation/leaf/non-empty.ts';
-export * from './builder/tree/tree-builder.ts';
+export * from './implementation/leaf/leaf-block.ts';
+export * from './implementation/leaf/leaf-tree.ts';
 export * from './builder/leaf/tree-builder.ts';
+export * from './builder/nonleaf/tree-builder.ts';
 
 export * from './implementation/nonleaf/nonleaf-tree.ts';
-
-export * from './builder/nonleaf/block-builder.ts';
-export * from './builder/nonleaf/tree-builder.ts';
 
 export * from './context.ts';
