+++ package.json
@@ -51,18 +51,18 @@
   },
   "devDependencies": {
     "@types/jest": "^27.0.2",
-    "@typescript-eslint/eslint-plugin": "^4.30.0",
-    "@typescript-eslint/parser": "^4.30.0",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
     "commitizen": "^4.2.4",
     "cz-conventional-changelog": "^3.3.0",
     "denoify": "^0.9.1",
     "docsify-cli": "^4.4.3",
-    "eslint": "^7.32.0",
+    "eslint": "^8.3.0",
     "husky": "^7.0.2",
     "jest": "^27.1.0",
     "lerna": "^4.0.0",
     "lint-staged": "^11.1.2",
-    "prettier": "^2.3.2",
+    "prettier": "^2.4.1",
     "rimraf": "^3.0.2",
     "ts-jest": "^27.0.7",
     "tsd": "^0.19.0",
