+++ packages/list/src/list-custom.ts
@@ -4,6 +4,7 @@ export * from './implementation/nonleaf/interface';
 export * from './implementation/nonleaf/nonleaf-block';
 export * from './implementation/tree/interface';
 export * from './implementation/tree/operations';
+export * from './implementation/leaf/non-empty';
 
 export * from './builder/block-builder';
 export * from './builder/builder-base';
@@ -11,15 +12,15 @@ export * from './builder/gen-builder';
 export * from './builder/leaf/block-builder';
 export * from './builder/leaf/leaf-builder';
 export * from './builder/nonleaf/nonleaf-builder';
+export * from './builder/tree/tree-builder';
+export * from './builder/nonleaf/block-builder';
 
 // circular dependencies
-export * from './implementation/leaf/non-empty';
-export * from './builder/tree/tree-builder';
+export * from './implementation/leaf/leaf-block';
+export * from './implementation/leaf/leaf-tree';
 export * from './builder/leaf/tree-builder';
+export * from './builder/nonleaf/tree-builder';
 
 export * from './implementation/nonleaf/nonleaf-tree';
-
-export * from './builder/nonleaf/block-builder';
-export * from './builder/nonleaf/tree-builder';
 
 export * from './context';
