+++ packages/list/test/nonleaf-block.test.ts
@@ -16,13 +16,17 @@ describe('NonLeafBlock', () => {
     const b4 = context.leafBlock([4, 5, 6]);
 
     {
-      const nl = context.nonLeafBlock(3, [b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(3, [b3], 1);
       const r = nl.append(b4) as NonLeafBlock<any, any>;
       expect(r.children).toBeArrayOf([b3, b4]);
       expect(r.length).toBe(6);
     }
     {
-      const nl = context.nonLeafBlock(12, [b3, b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        12,
+        [b3, b3, b3, b3],
+        1
+      );
       const r = nl.append(b4) as NonLeafTree<any, any>;
       expect(r.left).toBeInstanceOf(NonLeafBlock);
       expect(r.right).toBeInstanceOf(NonLeafBlock);
@@ -37,13 +41,17 @@ describe('NonLeafBlock', () => {
     const b4 = context.leafBlock([4, 5, 6]);
 
     {
-      const nl = context.nonLeafBlock(3, [b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(3, [b3], 1);
       const r = nl.appendInternal(b4);
       expect(r.children).toBeArrayOf([b3, b4]);
       expect(r.length).toBe(6);
     }
     {
-      const nl = context.nonLeafBlock(12, [b3, b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        12,
+        [b3, b3, b3, b3],
+        1
+      );
       const r = nl.appendInternal(b4);
       expect(r.children).toBeArrayOf([b3, b3, b3, b3, b4]);
       expect(r.length).toBe(15);
@@ -54,11 +62,15 @@ describe('NonLeafBlock', () => {
     const b3 = context.leafBlock([1, 2, 3]);
 
     {
-      const nl = context.nonLeafBlock(3, [b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(3, [b3], 1);
       expect(nl.canAddChild).toBe(true);
     }
     {
-      const nl = context.nonLeafBlock(12, [b3, b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        12,
+        [b3, b3, b3, b3],
+        1
+      );
       expect(nl.canAddChild).toBe(false);
     }
   });
@@ -67,11 +79,15 @@ describe('NonLeafBlock', () => {
     const b3 = context.leafBlock([1, 2, 3]);
 
     {
-      const nl = context.nonLeafBlock(3, [b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(3, [b3], 1);
       expect(nl.childrenInMax).toBe(true);
     }
     {
-      const nl = context.nonLeafBlock(12, [b3, b3, b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        12,
+        [b3, b3, b3, b3, b3],
+        1
+      );
       expect(nl.childrenInMax).toBe(false);
     }
   });
@@ -80,11 +96,15 @@ describe('NonLeafBlock', () => {
     const b3 = context.leafBlock([1, 2, 3]);
 
     {
-      const nl = context.nonLeafBlock(6, [b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        6,
+        [b3, b3],
+        1
+      );
       expect(nl.childrenInMin).toBe(true);
     }
     {
-      const nl = context.nonLeafBlock(3, [b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(3, [b3], 1);
       expect(nl.childrenInMin).toBe(false);
     }
   });
@@ -93,14 +113,22 @@ describe('NonLeafBlock', () => {
     const b3 = context.leafBlock([1, 2, 3]);
 
     {
-      const nl = context.nonLeafBlock(6, [b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        6,
+        [b3, b3],
+        1
+      );
       const r = nl.concat(nl) as NonLeafBlock<any, any>;
       expect(r).toBeInstanceOf(NonLeafBlock);
       expect(r.children).toBeArrayOf([b3, b3, b3, b3]);
       expect(r.length).toBe(12);
     }
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3],
+        1
+      );
       const r = nl.concat(nl) as NonLeafTree<any, any>;
       expect(r).toBeInstanceOf(NonLeafTree);
       expect(r.left).toBeInstanceOf(NonLeafBlock);
@@ -116,14 +144,22 @@ describe('NonLeafBlock', () => {
     const b3 = context.leafBlock([1, 2, 3]);
 
     {
-      const nl = context.nonLeafBlock(6, [b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        6,
+        [b3, b3],
+        1
+      );
       const r = nl.concatBlock(nl) as NonLeafBlock<any, any>;
       expect(r).toBeInstanceOf(NonLeafBlock);
       expect(r.children).toBeArrayOf([b3, b3, b3, b3]);
       expect(r.length).toBe(12);
     }
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3],
+        1
+      );
       const r = nl.concatBlock(nl) as NonLeafTree<any, any>;
       expect(r).toBeInstanceOf(NonLeafTree);
       expect(r.left).toBeInstanceOf(NonLeafBlock);
@@ -139,14 +175,22 @@ describe('NonLeafBlock', () => {
     const b3 = context.leafBlock([1, 2, 3]);
 
     {
-      const nl = context.nonLeafBlock(6, [b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        6,
+        [b3, b3],
+        1
+      );
       const r = nl.concatChildren(nl) as NonLeafBlock<any, any>;
       expect(r).toBeInstanceOf(NonLeafBlock);
       expect(r.children).toBeArrayOf([b3, b3, b3, b3]);
       expect(r.length).toBe(12);
     }
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3],
+        1
+      );
       const r = nl.concatChildren(nl);
       expect(r).toBeInstanceOf(NonLeafBlock);
       expect(r.children).toBeArrayOf([b3, b3, b3, b3, b3, b3]);
@@ -156,7 +200,7 @@ describe('NonLeafBlock', () => {
 
   it('context', () => {
     const b3 = context.leafBlock([1, 2, 3]);
-    const nl = context.nonLeafBlock(6, [b3, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(6, [b3, b3], 1);
     expect(nl.context).toBe(context);
   });
 
@@ -164,7 +208,11 @@ describe('NonLeafBlock', () => {
     const b1 = context.leafBlock([1, 2, 3]);
     const b2 = context.leafBlock([4, 5, 6]);
     const b3 = context.leafBlock([7, 8, 9]);
-    const nl = context.nonLeafBlock(9, [b1, b2, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+      9,
+      [b1, b2, b3],
+      1
+    );
 
     expect(nl.dropChildren(8)).toBeNull();
     expect(nl.dropChildren(0)).toBe(nl);
@@ -179,7 +227,11 @@ describe('NonLeafBlock', () => {
     const b3 = context.leafBlock([1, 2, 3]);
     const b3_2 = context.leafBlock([1, 2, 3]);
     {
-      const nl = context.nonLeafBlock(9, [b3_2, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3_2, b3, b3],
+        1
+      );
       const [r, u] = nl.dropFirst();
 
       expect(r).toBeInstanceOf(NonLeafBlock);
@@ -189,7 +241,7 @@ describe('NonLeafBlock', () => {
       expect(u).toBe(b3_2);
     }
     {
-      const nl = context.nonLeafBlock(3, [b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(3, [b3], 1);
       const [r, u] = nl.dropFirst();
 
       expect(r).toBeNull();
@@ -201,7 +253,11 @@ describe('NonLeafBlock', () => {
   it('dropInternal', () => {
     const b3 = context.leafBlock([1, 2, 3]);
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3],
+        1
+      );
       const [r, u, i] = nl.dropInternal(1);
       expect(r?.length).toBe(6);
       expect(u).toBeInstanceOf(LeafBlock);
@@ -209,7 +265,11 @@ describe('NonLeafBlock', () => {
       expect(i).toBe(1);
     }
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3],
+        1
+      );
       const [r, u, i] = nl.dropInternal(4);
       expect(r?.length).toBe(3);
       expect(u).toBeInstanceOf(LeafBlock);
@@ -217,7 +277,11 @@ describe('NonLeafBlock', () => {
       expect(i).toBe(1);
     }
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3],
+        1
+      );
       const [r, u, i] = nl.dropInternal(7);
       expect(r).toBeNull();
       expect(u).toBeInstanceOf(LeafBlock);
@@ -230,7 +294,11 @@ describe('NonLeafBlock', () => {
     const b3 = context.leafBlock([1, 2, 3]);
     const b3_2 = context.leafBlock([1, 2, 3]);
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3_2], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3_2],
+        1
+      );
       const [r, u] = nl.dropLast();
 
       expect(r).toBeInstanceOf(NonLeafBlock);
@@ -240,7 +308,7 @@ describe('NonLeafBlock', () => {
       expect(u).toBe(b3_2);
     }
     {
-      const nl = context.nonLeafBlock(3, [b3_2], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(3, [b3_2], 1);
       const [r, u] = nl.dropLast();
 
       expect(r).toBeNull();
@@ -251,7 +319,11 @@ describe('NonLeafBlock', () => {
 
   it('forEach', () => {
     const b3 = context.leafBlock([1, 2, 3]);
-    const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+      9,
+      [b3, b3, b3],
+      1
+    );
 
     const cb = jest.fn();
     nl.forEach(cb, TraverseState());
@@ -273,7 +345,11 @@ describe('NonLeafBlock', () => {
     const b1 = context.leafBlock([1, 2, 3]);
     const b2 = context.leafBlock([4, 5, 6]);
     const b3 = context.leafBlock([7, 8, 9]);
-    const nl = context.nonLeafBlock(9, [b1, b2, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+      9,
+      [b1, b2, b3],
+      1
+    );
 
     expect(nl.get(1)).toBe(2);
     expect(nl.get(4)).toBe(5);
@@ -284,7 +360,11 @@ describe('NonLeafBlock', () => {
     const b1 = context.leafBlock([1, 2, 3]);
     const b2 = context.leafBlock([4, 5, 6]);
     const b3 = context.leafBlock([7, 8, 9]);
-    const nl = context.nonLeafBlock(9, [b1, b2, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+      9,
+      [b1, b2, b3],
+      1
+    );
 
     expect(nl.getChild(0)).toBe(b1);
     expect(nl.getChild(1)).toBe(b2);
@@ -295,7 +375,11 @@ describe('NonLeafBlock', () => {
     {
       //  regular
       const b4 = context.leafBlock([1, 2, 3, 4]);
-      const nl = context.nonLeafBlock(8, [b4, b4], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        8,
+        [b4, b4],
+        1
+      );
 
       expect(nl.getCoordinates(0, false, false)).toEqual([0, 0]);
       // expect(nl.getCoordinates(0, true, false)).toEqual([0, 0]);
@@ -330,7 +414,11 @@ describe('NonLeafBlock', () => {
     {
       // not regular
       const b3 = context.leafBlock([1, 2, 3]);
-      const nl = context.nonLeafBlock(6, [b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        6,
+        [b3, b3],
+        1
+      );
 
       expect(nl.getCoordinates(0, false, false)).toEqual([0, 0]);
       expect(nl.getCoordinates(0, true, false)).toEqual([0, 0]);
@@ -361,14 +449,18 @@ describe('NonLeafBlock', () => {
 
   it('length', () => {
     const b3 = context.leafBlock([1, 2, 3]);
-    const nl = context.nonLeafBlock(6, [b3, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(6, [b3, b3], 1);
     expect(nl.length).toBe(6);
   });
 
   it('level', () => {
     const b3 = context.leafBlock([1, 2, 3]);
-    expect(context.nonLeafBlock(6, [b3, b3], 1).level).toBe(1);
-    expect(context.nonLeafBlock(6, [b3, b3], 2).level).toBe(2);
+    expect(
+      context.nonLeafBlock<number, LeafBlock<number>>(6, [b3, b3], 1).level
+    ).toBe(1);
+    expect(
+      context.nonLeafBlock<number, LeafBlock<number>>(6, [b3, b3], 2).level
+    ).toBe(2);
   });
 
   it('map', () => {
@@ -384,7 +476,7 @@ describe('NonLeafBlock', () => {
 
   it('nrChildren', () => {
     const b3 = context.leafBlock([1, 2, 3]);
-    const nl = context.nonLeafBlock(6, [b3, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(6, [b3, b3], 1);
     expect(nl.nrChildren).toBe(2);
   });
 
@@ -393,13 +485,17 @@ describe('NonLeafBlock', () => {
     const b4 = context.leafBlock([4, 5, 6]);
 
     {
-      const nl = context.nonLeafBlock(3, [b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(3, [b3], 1);
       const r = nl.prepend(b4) as NonLeafBlock<any, any>;
       expect(r.children).toBeArrayOf([b4, b3]);
       expect(r.length).toBe(6);
     }
     {
-      const nl = context.nonLeafBlock(12, [b3, b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        12,
+        [b3, b3, b3, b3],
+        1
+      );
       const r = nl.prepend(b4) as NonLeafTree<any, any>;
       expect(r.left).toBeInstanceOf(NonLeafBlock);
       expect(r.left.children[0]).toBe(b4);
@@ -414,13 +510,17 @@ describe('NonLeafBlock', () => {
     const b4 = context.leafBlock([4, 5, 6]);
 
     {
-      const nl = context.nonLeafBlock(3, [b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(3, [b3], 1);
       const r = nl.prependInternal(b4);
       expect(r.children).toBeArrayOf([b4, b3]);
       expect(r.length).toBe(6);
     }
     {
-      const nl = context.nonLeafBlock(12, [b3, b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        12,
+        [b3, b3, b3, b3],
+        1
+      );
       const r = nl.prependInternal(b4);
       expect(r.children).toBeArrayOf([b4, b3, b3, b3, b3]);
       expect(r.length).toBe(15);
@@ -432,7 +532,11 @@ describe('NonLeafBlock', () => {
     const b2 = context.leafBlock([4, 5, 6]);
     const b3 = context.leafBlock([7, 8, 9]);
 
-    const nl = context.nonLeafBlock(9, [b1, b2, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+      9,
+      [b1, b2, b3],
+      1
+    );
 
     const r = nl.reversed();
     expect(r.toArray()).toEqual([9, 8, 7, 6, 5, 4, 3, 2, 1]);
@@ -446,7 +550,11 @@ describe('NonLeafBlock', () => {
     const b2 = context.leafBlock([4, 5, 6]);
     const b3 = context.leafBlock([7, 8, 9]);
 
-    const nl = context.nonLeafBlock(9, [b1, b2, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+      9,
+      [b1, b2, b3],
+      1
+    );
 
     expect(nl.stream().toArray()).toEqual([1, 2, 3, 4, 5, 6, 7, 8, 9]);
     expect(nl.stream(true).toArray()).toEqual([9, 8, 7, 6, 5, 4, 3, 2, 1]);
@@ -457,7 +565,11 @@ describe('NonLeafBlock', () => {
     const b2 = context.leafBlock([4, 5, 6]);
     const b3 = context.leafBlock([7, 8, 9]);
 
-    const nl = context.nonLeafBlock(9, [b1, b2, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+      9,
+      [b1, b2, b3],
+      1
+    );
 
     expect(nl.streamRange({ amount: 10 }).toArray()).toEqual([
       1, 2, 3, 4, 5, 6, 7, 8, 9,
@@ -475,7 +587,11 @@ describe('NonLeafBlock', () => {
     const b1 = context.leafBlock([1, 2, 3]);
     const b2 = context.leafBlock([4, 5, 6]);
     const b3 = context.leafBlock([7, 8, 9]);
-    const nl = context.nonLeafBlock(9, [b1, b2, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+      9,
+      [b1, b2, b3],
+      1
+    );
 
     expect(nl.takeChildren(8)).toBe(nl);
     expect(nl.takeChildren(0)).toBeNull();
@@ -489,7 +605,11 @@ describe('NonLeafBlock', () => {
   it('takeInternal', () => {
     const b3 = context.leafBlock([1, 2, 3]);
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3],
+        1
+      );
       const [r, u, i] = nl.takeInternal(1);
       expect(r).toBeNull();
       expect(u).toBeInstanceOf(LeafBlock);
@@ -497,7 +617,11 @@ describe('NonLeafBlock', () => {
       expect(i).toBe(1);
     }
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3],
+        1
+      );
       const [r, u, i] = nl.takeInternal(4);
       expect(r?.length).toBe(3);
       expect(u).toBeInstanceOf(LeafBlock);
@@ -505,7 +629,11 @@ describe('NonLeafBlock', () => {
       expect(i).toBe(1);
     }
     {
-      const nl = context.nonLeafBlock(9, [b3, b3, b3], 1);
+      const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+        9,
+        [b3, b3, b3],
+        1
+      );
       const [r, u, i] = nl.takeInternal(7);
       expect(r?.length).toBe(6);
       expect(u).toBeInstanceOf(LeafBlock);
@@ -518,7 +646,11 @@ describe('NonLeafBlock', () => {
     const b1 = context.leafBlock([1, 2, 3]);
     const b2 = context.leafBlock([4, 5, 6]);
     const b3 = context.leafBlock([7, 8, 9]);
-    const nl = context.nonLeafBlock(9, [b1, b2, b3], 1);
+    const nl = context.nonLeafBlock<number, LeafBlock<number>>(
+      9,
+      [b1, b2, b3],
+      1
+    );
 
     expect(nl.toArray()).toEqual([1, 2, 3, 4, 5, 6, 7, 8, 9]);
     expect(nl.toArray({ amount: 0 })).toEqual([]);
